Make `create_schema_for_single_task` more flexible in terms of pydantic version
